Edit: Added logic for buying movies (Not working properly yet on postman)

+++ movie-rental-in/src/main/java/com/meawallet/usercrud/in/dto/CreateUserInRequest.java
@@ -1,7 +1,8 @@
 package com.meawallet.usercrud.in.dto;
 
 
+import java.math.BigDecimal;
 
-public record CreateUserInRequest(String name, Integer age, Integer credits) {
+public record CreateUserInRequest(String name, Integer age, BigDecimal credits) {
 
 }
